Refactor unnest chunk tests into dedicated module
Add a dedicated module for unnest chunking tests in
datafusion/core/tests/dataframe. Move relevant helper
functions and tests to a new focused file, ensuring
mod.rs remains cleaner. This organizes the test suite
for chunking into a more cohesive structure.
